// src/types/api/responses/client.response.ts
export interface ClientResponse {
    id: string;
    businessName: string;
    tradeName: string;
    address: string;
    email: string;
    phone: string;
    website: string;
    countryId: string;
    countryName: string;
    regionId: string;
    regionName: string;
    cityId: string;
    cityName: string;
    organizationId: string;
    organizationType: string;
    identifierType: string;
    identificationNumber: number;
    contactName: string;
    contactEmail: string;
    contactPhone: string;
}